requirements.txt
async-timeout==4.0.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
redis==5.1.0
Werkzeug==3.0.4
